Hackathon/Paulo/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
opencv-python>=4.5.0
pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.4.0
scikit-learn>=1.0.0

